Updating SfarScore to a double instead of an int.

+++ Messari/Models/Response/Assets/AssetProfileGeneralRegulation.cs
@@ -7,7 +7,7 @@ namespace Messari.Models.Response.Assets
         [JsonProperty("regulatory_details")]
         public string RegulatoryDetails { get; set; }
         [JsonProperty("sfar_score")]
-        public int SfarScore { get; set; }
+        public double SfarScore { get; set; }
         [JsonProperty("sfar_summary")]
         public string SfarSummary { get; set; }
     }
